Resolved crash when near and far is set before init. Moved SGCTWindow internal texture enums to private to prevent usage errors. Resolved warning in websocket example.

+++ src/apps/websocketsExample/UserData.cpp
@@ -40,7 +40,7 @@ void UserData::setCartesian2d(int x, int y, float r, float g, float b, float tim
     float r2 = s*s + t*t;
     if( r2 <= 1.0f )
     {
-        mPhi = sqrt(r2) * DEG90_TO_RAD;
+        mPhi = sqrtf(r2) * DEG90_TO_RAD;
         mTheta = atan2f(s,t);
     }
